chore: remove additional option file.
