--- pkg/plugins/gateway/base_http/option.go
@@ -1,35 +1,0 @@
-// Copyright 2021 Nitric Pty Ltd.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package base_http
-
-import (
-	"github.com/fasthttp/router"
-	"github.com/nitrictech/nitric/pkg/worker"
-	"github.com/valyala/fasthttp"
-)
-
-type BaseHttpGatewayOption = func(*BaseHttpGateway)
-
-func WithMiddleware(mw func(*fasthttp.RequestCtx, worker.WorkerPool) bool) BaseHttpGatewayOption {
-	return func(bhg *BaseHttpGateway) {
-		bhg.mw = mw
-	}
-}
-
-func WithRouter(routeRegister func(*router.Router, worker.WorkerPool)) BaseHttpGatewayOption {
-	return func(bhg *BaseHttpGateway) {
-		bhg.routeReg = routeRegister
-	}
-}
